Add tests for fast-forward lines surrounded by spaces
isFastForwardLine and parseFastForwardLimit currently don't match each
other in how they handle spaces on the outside of the fast-forward
markers.

isFastForwardLine allows them and will detect " >> 2 >> " as a
fast-forward line, but parseFastForwardLimit does not handle the spaces
correctly. It only trims the line that substring(int, int) is invoked on
and does not do so when determining the end index based on the string
length.
This can lead to an IndexOutOfBoundsException being thrown for strings
with three or more surrounding spaces, like "  >> 2 >> ". Strings with
one or two surround spaces, like " >> 2 >> ", are treated like an
unlimited fast-forward.

We add test cases that expect that spaces around fast-forward markers
are ignored, and such fast-forward lines are treated equally, such that
isFastForwardLine implies that parseFastForwardLimit will be successful
(unless the limit is not positive, in which case an exception is
thrown).