Treat text blocks as files in @CsvSource
PR #2721 introduced experimental support for text blocks in @CsvSource;
however, there was room for improvement.

Prior to this commit, a CSV record within a text block could not
contain a new line (\n), even if it was within a quoted string;
whereas, this is supported when using @CsvSource's value attribute. In
addition, comments do not make sense in a single string supplied via
@CsvSource's value attribute, but they do make sense within a text
block.

This commit refines the text block support in @CsvSource by treating
text blocks as complete CSV files, including support for comments
beginning with a `#` symbol as well as support for new lines within
quoted strings.

Closes #2734