Make quote character in @CsvSource configurable
Prior to this commit, the quote character for quoted strings in
@CsvSource was hard coded to a single quote (') and could not be
changed; however, with the recently added support for text blocks, it
may be desirable to change the quote character to something else.

This commit introduces a new quoteCharacter attribute in @CsvSource
that allows the user to change the quote character. The quoteCharacter
defaults to a single quote for backward compatibility.

Closes #2735