Align index for line in @CsvSource with index in display name
Prior to this commit, the index reported for an error in a CSV line
supplied via @CsvSource was 0-based; whereas, the index in the display
name for the CSV line (in the TestTemplateInvocationTestDescriptor
created by the TestTemplateTestDescriptor) was 1-based.

This commit aligns these sets of indexes to be 1-based to avoid
confusion.