Extend ParallelExecutionConfiguration to specify a saturated predicate
Add the method getSaturatePredicate to the interface ParallelExecutionConfiguration.
This allows a custom configuration to supply a predicate to handle the case
where the ForkJoinPool has reached the maximum pool size but all threads
are currently blocked.
Add a constructor to DefaultParallelExecutionConfiguration allowing to
speciy the predicate and modify the existing constructor to specify
null for the predicate.
Modify DefaultParallelExecutionConfigurationStrategyTests to check the
saturate predicate value in the default and custom configurations.

Issue: #2787