Do not pass the InternalTestPlan to TestExecutionListener
Hyrum's Law suggests that given a sufficient number of users of an API, any
observable behaviour will be depended on by someone. By not passing the
`InternalTestPlan` but rather the `TestPlan` we may still avoid the situation
where a test execution listener through reflection tries to access the
discovery result.