Add failIfNoTests attribute to `@Suite`
New `failIfNoTests` attribute added to `@Suite`. This will fail the
suite if no tests were discovered.

This mechanism is implemented by having the `SuiteLauncher` listen
in on the test execution while also delegating any execution events
to the engine execution listener of the parent engine. If the
listener did not hear any test executions the suite will mark its
own execution as a failure.

To prevent pruning the `SuiteTestDescriptor` from being pruned
prematurely and not being executed at all, the suite  describes
itself a as test container that may register test. In practice
the container will never register any tests.

The `InternalTestPlan` is not passed to `TestExecutionListener` because
Hyrum's Law suggests that given a sufficient number of users of an API, any
observable behaviour will be depended on by someone. By not passing the
`InternalTestPlan` but rather the `TestPlan` we may still avoid the situation
where a test execution listener through reflection tries to access the
discovery result.

Closes #2823