Use user's console charset if available
Prior to this commit the charset returned by
`Charset.defaultCharset()` was always used to
determine the default theme used by the JUnit
Console Launcher application.

Now, on Java 17 and later, Java's Console API is
queried in order to retrieve the user-specified
of the charset used by the system console --
before falling back to the default charset.