Update DefaultParallelExecutionConfigurationStrategy.java
Since Turkish has a special characters as "ı" and "İ", uppercase convertion here causes all test methods to finish with exit code -1. Because uppercase of "i" is "İ" in Turkish and that breaks the code. "dynamic" > "DYNAMİC"


There is a special comment for Turkish characters on reference documentation of java.lang.String:

https://docs.oracle.com/javase/7/docs/api/java/lang/String.html#toUpperCase(java.util.Locale)