Add thread mode to @Timeout (#2949)
Timeouts can now be applied using one of the following three thread
modes: `SAME_THREAD`, `SEPARATE_THREAD`, or `INFERRED`.

When `SAME_THREAD` is used, the execution of the annotated method
proceeds in the main thread of the test. If the timeout is exceeded,
the main thread is interrupted from another thread. This is done to
ensure interoperability with frameworks such as Spring that make use of
mechanisms that are sensitive to the currently running thread — for
example, `ThreadLocal` transaction management.

On the contrary when `SEPARATE_THREAD` is used, like the
`assertTimeoutPreemptively()` assertion, the execution of the annotated
method proceeds in a separate thread, this can lead to undesirable side
effects, see <<writing-tests-assertions-preemptive-timeouts>>.

When `INFERRED` (default) thread mode is used, the thread mode is
resolved via the `junit.jupiter.execution.timeout.thread.mode.default`
configuration parameter. If the provided configuration parameter is
invalid or not present then `SAME_THREAD` is used as fallback.

Resolves #2087.

Co-authored-by: Marc Philipp <mail@marcphilipp.de>