Revise fix for @MethodSource factory method resolution
This commit revises the fix for the @MethodSource factory method
resolution regression by also ignoring candidate factory methods (i.e.,
those with the same name as the designated factory method) with an
invalid return type. Such methods may in fact be local utility methods
that are not intended to be used as a factory method.

Closes #3001