Fix header support in @CsvSource for records supplied via value attribute
Previously, when parsing the `value` attribute of the `@CsvSource`
annotation with `useHeadersInDisplayName` set to true,
`CsvArgumentsProvider` didn't handle headers correctly. After parsing
the first line with headers, instead of moving to the next line, it
tried to validate a not-yet-parsed CSV record which lead to an
exception being thrown. This was due to the missing `continue`
statement in the `for` loop.

Now, with `continue` added, as soon as headers are parsed, the `for`
loop is advanced to the next iteration and to the next line of CSV
input, containing the expected CSV record.

Fixes #3010
Closes #3011