Fix fallback to ForkJoinPools Java 7 constructor on Java 9 and higher
When an invalid ParallelExecutionConfiguration is provided on Java 9+ the
`ForkJoinPoolHierarchicalTestExecutorService` will fall back to using the
since Java 7 constructor. As a result misconfiguration goes unnoticed.

Separated the exceptions thrown by finding the Java 9+ constructor from those
thrown by invoking the constructor with an invalid configuration. The former
can be ignored, the latter should not be.

Fixes #3045.