Use URLEncode for all characters in UniqueId segment values
Previous to this commit, only a minimal subset of characters has been escaped,
thus leaving the UniqueId unsafe for use in an url.
With the addition of a string representation of selectors based on the URI
format, it has become necessary to make UniqueId url-safe.
Instead of double-encoding the UniqueId, we make the default `toString()` form URL safe.

Still some problems regarding the limits of encoding/decoding