Model parameters using org.gradlex.build-parameters
Introduces the org.gradlex.build-parameters plugin to model parameters
that can be supplied to this build. For this the existing build logic
in gradle/plugins is moved to a submodule called 'common' and a new
module called 'build-parameters' is introduced. This way, build logic
can access build parameters using the `buildParameters` pre-compiled
accessor in Kotlin.

Occurences of project.findProperty as well as project.hasProperty
are replaced by properly modeled build parameters. Since validation
of parameters is not turned off, the build-parameters plugin will
fail the build is a parameter is supplied which is not modeled.
For that reason, any other parameters, such as signing or deployment
parameters that are supplied by developers only during releasing need to
be added before this can be merged.