Model parameters using org.gradlex.build-parameters
Introduces the org.gradlex.build-parameters plugin to model parameters
that can be supplied to this build. For that an included build is added
that contains the parameter configurations. It's necessary to put this
into an included build instead of buildSrc because buildSrc can not
build plugins that can be applied to the settings script.

The details on how to conncet to the remote build cache as well as the
enablement of test distribution are then read from the
BuildParametersExtension in settings.gradle.kts. Injection of the
parameters on CI now uses auto-generated environment variable look up
provided by the build parameters plugin.

The next step would be to also model parameters used in other parts of
the build logic using the build parameters plugin. This however requires
a larger refactoring of buildSrc because buildSrc can only depend on
included build modules that are included via Settings.includeBuild, but
we need to include the build-logic build via
pluginManagement.includeBuild due to the reasons stated in the
beginning.