Introduce dry-run execution mode (#3165)
The new `junit.platform.execution.dryRun.enabled` configuration
parameter can be used to enable dry-run mode in which tests returned
from test discovery are pretended to be executed and test
started/finished events are reported to registered
`TestExecutionListeners`. This can be useful to test changes in the
configuration of a build or to verify a listener is called as expected
without having to wait for all tests to be executed.