Use ClassLoader from Class in ReflectionUtils.loadRequiredParameterType()
Prior to this commit, the loadRequiredParameterType() method in
ReflectionUtils used tryToLoad(String) to load the class for the
specified parameter type name via the "default" ClassLoader. For
standard ClassLoader arrangements this works fine; however, for more
complex ClassLoader arrangements – such as those used in OSGi – loading
the parameter type via the default ClassLoader may fail.

For example, DiscoverySelectors.selectMethod(Class<?>, String, String)
would previously fail to select the specified method if the parameter
types could not be loaded using the default ClassLoader.

To address that issue, this commit modifies loadRequiredParameterType()
so that it uses the tryToLoad(String, ClassLoader) variant, specifying
a ClassLoader that is more likely to be able to load the parameter type.

Specifically, the ClassLoader for the supplied Class is used.
Alternatively, if that ClassLoader is null (for example, because the
Class represents a primitive type or void), then
ClassLoaderUtils.getDefaultClassLoader() is used to obtain the
"default" ClassLoader (which is effectively a fallback to the original
behavior).

Co-authored-by: kriegfrj <fr.jkrieg@greekwelfaresa.org.au>
Closes #2104