Add `junit.jupiter.tempdir.factory.default` configuration property
You can use the `junit.jupiter.tempdir.factory.default` configuration
parameter to specify the fully qualified class name of the
`TempDirFactory` you would like to use by default. Just like for
factories configured via the `factory` attribute of the `@TempDir`
annotation, the supplied class has to implement the `TempDirFactory`
interface. The default factory will be used for all `@TempDir`
annotations unless the `factory` attribute of the annotation specifies
a different factory.

In summary, the factory for a temporary directory is determined
according to the following precedence rules:

1. The `factory` attribute of the `@TempDir` annotation, if present
2. The default `TempDirFactory` configured via the configuration
   parameter, if present
3. Otherwise, `org.junit.jupiter.api.io.TempDirFactory$Standard` will be
   used.

Co-authored-by: Marc Philipp <mail@marcphilipp.de>
Co-authored-by: Sam Brannen <sam@sambrannen.com>