Allowing selecting methods in test suites
This commit adds two annotations to the ‘junit-platform-suite-api’,
and also includes three types of tests to ensure that the annotations
functionality is working correctly.

The two annotations allow the user to select methods in suite testing
based on the name, the class where the method is located (optional)
and the input parameters (optional).

The three added tests test the operation of the two added annotations,
examining three possible situations for using the added annotations.

Resolves : #3012