Support method selection with var-args parameter types in DiscoverySelectors
Prior to this commit, when selecting a method or nested method in
DiscoverySelectors, the parameter list could only be supplied as a
comma-delimited string -- for example, "int,java.lang.String,boolean[]".

This commit introduces selectMethod() and selectNestedMethod() variants
in DiscoverySelectors that accept a `Class<?>...` (var-args) argument
of parameter types as a type-safe alternative to providing the
names of parameter types as a comma-delimited string.

Closes #3326
Closes #3327