Allow selecting classes by name via `@SelectClasses`
This commit introduces a `String[] names` attribute in `@SelectClasses`
to select classes by fully-qualified class name.

This new attribute can be used in conjunction with or instead of
`value`.

The intent is to support selecting classes that are not visible to the
class in which `@SelectClasses` is declared -- for example a `@Suite`
class.

Resolves #3325.