Add separate dependency-submission workflow
With the introduction of `gradle/actions/dependency-submission`, it is now
simpler (and recommended) to use a separate workflow for generation and
submission of GitHub Dependency Graph.

This workflow attempts to detect and submit all dependencies that would
be resolved during build execution, without requiring the execution of any
particular task. In basic testing it appears that the generated dependency
graph contains the same dependencies as before.

A few things to note:
The new workflow will submit a dependency graph with a different "correlator" to
the previous one. This means that duplicate dependencies (and alerts) may appear
until the older graph ages out and is automatically purged. (Period of hours to days).

Manually dismissed Dependabot Alerts may need to be re-dismissed after
switching to the new workflow.