Invoke @⁠AutoClose methods via interfaces whenever possible
Prior to this commit, the @⁠AutoClose extension attempted to invoke a
"close" method via the most specific "close" method found in the type
hierarchy. However, this resulted in an InaccessibleObjectException if
the "close" method was implemented in a non-public type from a
different module (such as from the java.base module).

In order to avoid issues with the enforcement of strict encapsulation
within the JVM, this commit attempts to always invoke a "close" method
via the declaring interface.

For a "close" method in a type that implements AutoCloseable, we now
simply cast the object to AutoCloseable and invoke the "close()" method
directly, without using reflection.

For any other method, we use the new getInterfaceMethodIfPossible()
utility in ReflectionUtils to locate an interface method (if such a
method exists) and invoke that.

The latter was graciously donated by @jhoeller from the Spring
Framework.

Closes #3684