Apply suite configuration to tests selected by unique ID (#3694)
Tests in a suite selected by unique id did not receive their
configuration because this was only applied to the discovery request
builder when a suite class was selected. By splitting the processing of
configuration annotations and select and filter annotations, the
configuration can now always be applied prior to discovery.

Fixes #3693.