Consistently throw NamespacedHierarchicalStoreException
NamespacedHierarchicalStore previously threw an IllegalStateException
if the store was accessed or modified after it had been closed;
however, we typically aim to throw an instance of JUnitException
whenever we intentionally throw an exception.

With this commit, NamespacedHierarchicalStore now consistently throws
NamespacedHierarchicalStoreException (which is a subclass of
JUnitException) instead of IllegalStateException.

See #3614