Introduce named argument set support for @⁠ParameterizedTest
Prior to this commit, it was possible to use the Named API to provide
custom names for individual arguments for a @⁠ParameterizedTest when
using @⁠MethodSource, @⁠FieldSource, or @⁠ArgumentsSource, but it was
not possible to provide a custom name for an entire set of arguments
without resolving to obtuse workarounds.

To address that, this commit introduces an argumentSet(String,Object...)
static factory method in Arguments for providing a custom name for an
entire set of arguments for a @⁠ParameterizedTest.

This commit also introduces the following placeholders for the display
name of a @⁠ParameterizedTest.

- {argumentSetName} :: the name of the argument set

- {argumentSetNameOrArgumentsWithNames} :: resolves to {argumentSetName}
  or {argumentsWithNames}, depending on how the arguments are supplied

In addition, the default display name for a @⁠ParameterizedTest
is now "[{index}] {argumentSetNameOrArgumentsWithNames}".

For example, given the following...

@⁠ParameterizedTest
@⁠FieldSource
void argumentSets(File file1, File file2) {
}

static List<Arguments> argumentSets = List.of(
    argumentSet("Important files", new File("path1"), new File("path2")),
    argumentSet("Other files", new File("path3"), new File("path4"))
);

... the resulting display names are:

[1] Important files
[2] Other files

Closes #3818