Process extensions registered on non-static fields earlier (#3895)
This PR changes extensions declared via @ExtendWith on non-static fields
to be registered before the test instance is instantiated. Since these 
declarative extensions are documented to be sorted along with 
programmatic extensions registered via @RegisterExtension the latter 
are also registered earlier. Since, however, non-static fields are not 
initialized until the test class is instantiated, this PR introduces the
internal concept of "uninitialized extensions" that are registered 
early and initialized whenever an instance of the test class is created.

Fixes #3437. Resolves #3463.

Co-authored-by: Sam Brannen <104798+sbrannen@users.noreply.github.com>