Support arbitrary Java versions with JRE conditions
Prior to JUnit Jupiter 5.12, JRE-based conditions could only rely on
predefined constants in the JRE enum. Furthermore, those constants are
only updated as long as the particular JUnit Jupiter branch is
supported. For example, once JUnit Jupiter 5.12 is released, there is
no guarantee that the JRE enum constants will be updated in the 5.11.x
branch.

Consequently, users previously did not have the ability to enable or
disable tests for Java versions released after a particular JUnit
Jupiter branch was no longer supported.

To address that, this commit introduces support for arbitrary Java
versions in the JRE enum and related condition annotations.

Users can now specify arbitrary Java versions via the `versions`
attributes in @⁠EnabledOnJre and @⁠DisabledOnJre and via the
`minVersion` and `maxVersion` attributes in @⁠EnabledForJreRange and
@⁠DisabledForJreRange.

Closes: #3930
Closes: #3931