Fix potential double acquisition in `SingleLock`
`SingleLock` uses a `ForkJoinPool.ManagedBlocker` for efficient blocking
in a fork-join pool. In the `isReleasable()` method it calls `tryLock()`
without remembering the outcome. The `block` method also unconditionally
called `lockInterruptibly()` even if the lock was already acquired.
The problem that can arise is that the lock is acquired multiple times,
but only released once.

`CompositeLockManagedBlocker` is also updated, so that `block()` checks
for `acquired` before locking.