Fix deadlock in `ForkJoinPoolHierarchicalTestExecutorService` (#3981)
The service now checks if the `ExclusiveTask` that should run
is executed on a thread that is already executing another task.
If this is scenario is detected, it checks if the lock is compatible to
the enclosing locks.
1. If compatible, it is executed and marked done
2. If incompatible, it is added to a list of deferred tasks and left
   unfinished. The deferred tasks will be re-forked afterwards.

Fixes #3945.

---------

Co-authored-by: Marc Philipp <mail@marcphilipp.de>
(cherry picked from commit 88f88591abc9900a3e30f0c3a9fca524b5c0f8b5)