Allow work stealing when more locks can be acquired without deadlock
This allows tasks with only read locks to be stolen by threads that are
currently holding only read locks. The new compatibility check would
also support write locks but `NodeTreeWalker` forces children of tasks
with write locks to be executed on the same thread and thus no work is
being stolen.