Improve performance of ParameterizedTestSpiInstantiator
Since Class.getDeclaredConstructors() clones the constructors array AND
makes "child copies" of the constructors, it's usually better to avoid
repeated calls to getDeclaredConstructors() for a single use case. In
addition, it's good to avoid the use of Optional, multiple Streams, and
try-catch blocks if feasible.

This commit reworks ParameterizedTestSpiInstantiator in order to
achieve that by switching to old-school Java constructs like arrays,
for-loops, and if-blocks.

See #4018
See #4025