Add JUnit-specific metadata to open-test-reporting's HTML report (#4116)
When creating the event-based XML report, the
`OpenTestReportGeneratingListener` includes JUnit-specific metadata:
type, unique ID, and legacy reporting name. This PR implements 
open-test-reporting's `Contributor` SPI to include that data in the new
HTML report of open-test-reporting.

Moreover, it configures the build to create a single report for all test
tasks per subproject and uploads it as part of GitHub Action runs.