Pass ServiceLoader to Utils to avoid failing its caller check
When used on the module path, `ServiceLoader` determines its calling
module from the stack and check whether its module descriptor contains a
`uses` directive for the service type. If not, it will throw an
exception. Therefore, `ServiceLoader.load` needs to be called directly
in the module using the service but the filtering can still occur in
junit-platform-commons.