Allow attaching files to test results (#4138)
Being able to attach files such as screenshots or extra log files to a 
test is useful to diagnose the outcome of tests. This PR adds an API for
Jupiter test authors to do so (`TestReporter.publishFile`) and includes
it when writing the Open Test Reporting XML output (in
`OpenTestReportGeneratingListener`) via a new method on 
`TestExecutionListener`. Moreover, it adds `OutputDirectoryProvider` to 
`EngineDiscoveryRequest` so other engines can also attach files and
write them to the same output directory and makes it available to 
`TestExecutionListener` implementations via `TestPlan`.

The default location of the XML output is changed from
`OUTPUT_DIR/junit-platform-events-*.xml` to 
`OUTPUT_DIR/open-test-report.xml`. The output directory can be made
unique by using the `{uniqueNumber}` placeholder in the
`junit.platform.reporting.output.dir` configuration parameter.