Attribute non-test thread output to most recent test thread (#4200)
When using frameworks or running external processes, test output is
often written on other threads than the test. When tests are executed
sequentially that output can be attributed unambiguously to the current
test. If tests are run in parallel, picking the right test to attribute
output to becomes much harder, though.