Add native-image.properties files to jars (#4208)
This allows adding or removing entries from the set of classes that are 
initialized at native image build time without requiring changes to 
graalvm/native-build-tools. The new precompiled script plugin introduced
in this commit defines two Gradle tasks: `nativeImageProperties` and 
`validateNativeImageProperties`. The former is used for writing the 
`native-image.properties` file to the correct resource folder based (by
convention) on the `src/nativeImage/initialize-at-build-time` file in 
each affected subproject. The latter validates that all listed classes
exist and acts as a safeguard when classes are renamed or deleted.

Resolves #4207.